# [Enhancement] Refine mem tracker mechanism to support operator level tracing