--- a/be/src/exec/sorted_streaming_aggregator.cpp
+++ b/be/src/exec/sorted_streaming_aggregator.cpp
@@ -269,9 +269,8 @@ SortedStreamingAggregator::~SortedStreamingAggregator() {
     }
 }
 
-Status SortedStreamingAggregator::prepare(RuntimeState* state, ObjectPool* pool, RuntimeProfile* runtime_profile,
-                                          MemTracker* mem_tracker) {
-    RETURN_IF_ERROR(Aggregator::prepare(state, pool, runtime_profile, mem_tracker));
+Status SortedStreamingAggregator::prepare(RuntimeState* state, ObjectPool* pool, RuntimeProfile* runtime_profile) {
+    RETURN_IF_ERROR(Aggregator::prepare(state, pool, runtime_profile));
     _streaming_state_allocator =
             std::make_shared<StateAllocator>(_mem_pool.get(), _state->chunk_size(), _agg_states_total_size);
     return Status::OK();