--- a/be/src/exec/sorted_streaming_aggregator.h
+++ b/be/src/exec/sorted_streaming_aggregator.h
@@ -27,8 +27,7 @@ public:
     SortedStreamingAggregator(AggregatorParamsPtr&& params);
     ~SortedStreamingAggregator() override;
 
-    Status prepare(RuntimeState* state, ObjectPool* pool, RuntimeProfile* runtime_profile,
-                   MemTracker* mem_tracker) override;
+    Status prepare(RuntimeState* state, ObjectPool* pool, RuntimeProfile* runtime_profile) override;
     Status streaming_compute_agg_state(size_t chunk_size);
 
     StatusOr<ChunkPtr> pull_eos_chunk();