--- a/be/src/exec/stream/aggregate/stream_aggregator.h
+++ b/be/src/exec/stream/aggregate/stream_aggregator.h
@@ -65,9 +65,8 @@ public:
 
     Status open(RuntimeState* state) { return Aggregator::open(state); }
 
-    Status prepare(RuntimeState* state, ObjectPool* pool, RuntimeProfile* runtime_profile,
-                   MemTracker* mem_tracker) override {
-        RETURN_IF_ERROR(Aggregator::prepare(state, pool, runtime_profile, mem_tracker));
+    Status prepare(RuntimeState* state, ObjectPool* pool, RuntimeProfile* runtime_profile) override {
+        RETURN_IF_ERROR(Aggregator::prepare(state, pool, runtime_profile));
         RETURN_IF_ERROR(_prepare_state_tables(state));
         return Status::OK();
     }