… be/src/runtime/current_thread.h
… be/src/runtime/mem_tracker.cpp
… be/src/runtime/mem_tracker.h
… be/src/runtime/runtime_state.cpp
… be/src/util/runtime_profile.h
… be/test/exec/stream/stream_aggregator_test.cpp
… fe/fe-core/src/main/java/com/starrocks/common/util/Counter.java
… fe/fe-core/src/main/java/com/starrocks/common/util/ProfileManager.java
… fe/fe-core/src/main/java/com/starrocks/load/loadv2/LoadLoadingTask.java
… fe/fe-core/src/main/java/com/starrocks/qe/Coordinator.java
… fe/fe-core/src/main/java/com/starrocks/qe/StmtExecutor.java