Signed-off-by: liuyehcf <1559500551@qq.com>

## What type of PR is this：
- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Support operator level memory tracing

Currently, we only support instance level memory tracing. And we can further support operator level memory tracing by 

1. Attach Operator's MemTracker, as child, to FragmentInstance's MemTracker
2. Set thread local mem_tracker when calling `push_chunk`、`pull_chunk`、`set_finishing`、`set_finished`、`set_cancelled` and `close`. Memory consumption information will be recorded both in Operator's MemTracker and FragmentInstance's MemTracker.

## Improve MemTracker mechanism

For now, we only record the peak memory usage of a given query, and it is not enough when it comes to the requirements of the memory test framework, which demands a query should have stable memory usage.

So we need to record another two metrics called `MemTracker::_allocation` and `MemTracker::_deallocation`, which respectively record allocation and deallocation.

### Be compatible with CurrentThread

For sake of performance, not every memory allocation and deallocation will be passed to MemTracker for updation. CurrentThread will first cache it, when exceed the threshold `CurrentThread::BATCH_SIZE`, then update MemTracker. This cache processing will violate the correctness of the `MemTracker::_allocation` and `MemTracker::_deallocation`. 

So we  need to improve the cache mechanism as well by adding two field `CurrentThread::_allocated_cache_size`  and `CurrentThread::_deallocated_cache_size` 

## Metrics renames

Rename:

1. QueryCpuTime -> QueryCumulativeCpuTime
2. QueryMemUsage -> MachinePeakMemoryUsage
3. OperatorCumulativeTime -> QueryCumulativeOperatorTime

## Example

After this pr
* every Operator's level has `OperatorAllocatedMemoryUsage/OperatorDeallocatedMemoryUsage`
    * We hide `OperatorPeakMemoryUsage`, because in some workflow, suppose A and B are two adjcent operators, memory only allocated in A and only deallocated in B, the peak memory usage for A and B is meaningless.
* every Instance's level has `InstanceAllocatedMemoryUsage/InstanceDeallocatedMemoryUsage/InstancePeakMemoryUsage`
* every Query's level has `QueryAllocatedMemoryUsage/QueryDeallocatedMemoryUsage/MachinePeakMemoryUsage`

```
...
  Execution Profile 666250bb-a39f-11ed-9724-00163e1f2aee:
     - QueryCumulativeCpuTime: 2m47s
     - MachinePeakMemoryUsage: 12.722GB
     - ExecutionWallTime: 8s82ms
     - QueryAllocatedMemoryUsage: 220.89 GB
     - QueryCumulativeOperatorTime: 10s753ms
 ...
    Fragment 2:
       - InstanceAllocatedMemoryUsage: 63.31 GB
         - __MAX_OF_InstanceAllocatedMemoryUsage: 21.35 GB
         - __MIN_OF_InstanceAllocatedMemoryUsage: 20.63 GB
       - InstancePeakMemoryUsage: 203.69 MB
         - __MAX_OF_InstancePeakMemoryUsage: 203.98 MB
         - __MIN_OF_InstancePeakMemoryUsage: 203.44 MB
       - QueryMemoryLimit: 38.07 GB
 ...
        EXCHANGE_SINK (plan_node_id=1):
          CommonMetrics:
             - CloseTime: 3.8us
               - __MAX_OF_CloseTime: 19.366us
               - __MIN_OF_CloseTime: 447ns
             - OperatorAllocatedMemoryUsage: 19.00 KB
               - __MAX_OF_OperatorAllocatedMemoryUsage: 4.31 KB
               - __MIN_OF_OperatorAllocatedMemoryUsage: 640.00 B
             - OperatorDeallocatedMemoryUsage: 6.02 KB
               - __MAX_OF_OperatorDeallocatedMemoryUsage: 1.55 KB
               - __MIN_OF_OperatorDeallocatedMemoryUsage: 64.00 B
...
```

## Performance test

operator level memory tracing may introduce some overhead from 2 aspects:

1. thread_local obj `tls_mem_tracker` will be switched
2. every time recording memory allocation/dellocation, before this pr, two mem_tracker will be updated(Query level and Instance level), and after this pr, there will be one more Operator mem_tracker

So I have conduct a experiment to find out whether this overhead can lead to significant performance deduction. And the results show that the overhead is negligible

TPCH-100g, 16c/64g * 3

| Query | Before | After |
|:--|:--|:--|
| Q1 | 1.978 | 1.98 |
| Q2 | 0.156 | 0.135 |
| Q3 | 0.78 | 0.758 |
| Q4 | 0.446 | 0.445 |
| Q5 | 1.308 | 1.309 |
| Q6 | 0.06 | 0.063 |
| Q7 | 1.083 | 1.105 |
| Q8 | 0.594 | 0.606 |
| Q9 | 2.032 | 2.067 |
| Q10 | 0.528 | 0.523 |
| Q11 | 0.191 | 0.184 |
| Q12 | 0.192 | 0.192 |
| Q13 | 1.892 | 1.891 |
| Q14 | 0.168 | 0.17 |
| Q15 | 0.31 | 0.273 |
| Q16 | 0.199 | 0.201 |
| Q17 | 0.376 | 0.379 |
| Q18 | 4.163 | 4.149 |
| Q19 | 0.473 | 0.479 |
| Q20 | 0.269 | 0.267 |
| Q21 | 0.082 | 0.096 |
| Q22 | 0.394 | 0.374 |

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be auto backported to target branch
  - [ ] 2.5
  - [ ] 2.4
  - [ ] 2.3
  - [ ] 2.2