@wanpengfei-git 2023-02-02
#### [FE PR Coverage Check]

:disappointed: **fail** : 1 / 21 (04.76%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/qe/Coordinator.java|0|18|00.00%|[1598, 1603, 1606, 1607, 1608, 1630, 1634, 1639, 1644, 1645, 1646, 1647, 1648, 1649, 1652, 1653, 1654, 1655]|
|:large_blue_circle:|com/starrocks/qe/StmtExecutor.java|0|1|00.00%|[272]|
|:large_blue_circle:|com/starrocks/load/loadv2/LoadLoadingTask.java|0|1|00.00%|[234]|
|:large_blue_circle:|com/starrocks/common/util/Counter.java|1|1|100.00%|[]|