@liuyehcf 2023-02-02
run starrocks_admit_test