@liuyehcf 2023-02-02
run starrocks_fe_unittest