@liuyehcf 2023-02-06
run starrocks_admit_test