@wanpengfei-git 2023-02-07
#### [FE PR Coverage Check]

:disappointed: **fail** : 1 / 28 (03.57%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/qe/Coordinator.java|0|25|00.00%|[1598, 1599, 1604, 1607, 1608, 1609, 1611, 1612, 1613, 1635, 1639, 1644, 1649, 1650, 1651, 1652, 1653, 1654, 1655, 1656, 1657, 1660, 1661, 1662, 1663]|
|:large_blue_circle:|com/starrocks/qe/StmtExecutor.java|0|1|00.00%|[274]|
|:large_blue_circle:|com/starrocks/load/loadv2/LoadLoadingTask.java|0|1|00.00%|[234]|
|:large_blue_circle:|com/starrocks/common/util/Counter.java|1|1|100.00%|[]|