@liuyehcf 2023-02-07
run starrocks_fe_unittest