@liuyehcf 2023-02-07
run starrocks_admit_test