@wanpengfei-git 2023-02-08
run starrocks_admit_test